move original constants to the placeholder constant class
