--- src/main/java/uk/gov/hmcts/reform/sscs/config/PlaceholderConstants.java
@@ -1,30 +1,0 @@
-package uk.gov.hmcts.reform.sscs.config;
-
-public class PlaceholderConstants {
-
-    public static final String CASE_CREATED_DATE_LITERAL = "case_created_date";
-    public static final String REGIONAL_OFFICE_ADDRESS_LINE1_LITERAL = "rpc_address_line1";
-    public static final String REGIONAL_OFFICE_ADDRESS_LINE2_LITERAL = "rpc_address_line2";
-    public static final String REGIONAL_OFFICE_ADDRESS_LINE3_LITERAL = "rpc_address_line3";
-    public static final String REGIONAL_OFFICE_ADDRESS_LINE4_LITERAL = "rpc_address_line4";
-    public static final String REGIONAL_OFFICE_COUNTY_LITERAL = "rpc_address_county";
-    public static final String REGIONAL_OFFICE_PHONE_LITERAL = "rpc_address_phone";
-    public static final String REGIONAL_OFFICE_FAX_LITERAL = "rpc_address_fax";
-    public static final String REGIONAL_OFFICE_POSTCODE_LITERAL = "rpc_address_postcode";
-    public static final String DWP_ADDRESS_LINE1_LITERAL = "dwp_address_line1";
-    public static final String DWP_ADDRESS_LINE2_LITERAL = "dwp_address_line2";
-    public static final String DWP_ADDRESS_LINE3_LITERAL = "dwp_address_line3";
-    public static final String DWP_ADDRESS_LINE4_LITERAL = "dwp_address_line4";
-    public static final String BENEFIT_TYPE_LITERAL = "benefit_type";
-    public static final String APPELLANT_FULL_NAME_LITERAL = "appellant_full_name";
-    public static final String CASE_ID_LITERAL = "case_id";
-    public static final String NINO_LITERAL = "nino";
-    public static final String GENERATED_DATE_LITERAL = "generated_date";
-    public static final String SSCS_URL_LITERAL = "sscs_url";
-    public static final String SSCS_URL = "http://www.tribunals.gov.uk/";
-
-
-    private PlaceholderConstants() {
-        //
-    }
-}
