+++ src/main/java/uk/gov/hmcts/reform/sscs/service/placeholders/CommonPlaceholderService.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.config.PlaceholderConstants;
 import uk.gov.hmcts.reform.sscs.docmosis.config.PdfDocumentConfig;
 
 @Service
