+++ src/main/java/uk/gov/hmcts/reform/sscs/service/placeholders/PlaceholderConstants.java
@@ -1,0 +1,36 @@
+package uk.gov.hmcts.reform.sscs.service.placeholders;
+
+public final class PlaceholderConstants {
+
+    public static final String CASE_CREATED_DATE_LITERAL = "case_created_date";
+    public static final String BENEFIT_TYPE_LITERAL = "benefit_type";
+    public static final String APPELLANT_FULL_NAME_LITERAL = "appellant_full_name";
+    public static final String CASE_ID_LITERAL = "case_id";
+    public static final String NINO_LITERAL = "nino";
+    public static final String GENERATED_DATE_LITERAL = "generated_date";
+    public static final String SSCS_URL_LITERAL = "sscs_url";
+    static final String SSCS_URL = "http://www.tribunals.gov.uk/";
+
+    public static final String REGIONAL_OFFICE_ADDRESS_LINE1_LITERAL = "rpc_address_line1";
+    public static final String REGIONAL_OFFICE_ADDRESS_LINE2_LITERAL = "rpc_address_line2";
+    public static final String REGIONAL_OFFICE_ADDRESS_LINE3_LITERAL = "rpc_address_line3";
+    public static final String REGIONAL_OFFICE_ADDRESS_LINE4_LITERAL = "rpc_address_line4";
+    public static final String REGIONAL_OFFICE_COUNTY_LITERAL = "rpc_address_county";
+    static final String REGIONAL_OFFICE_PHONE_LITERAL = "rpc_address_phone";
+    static final String REGIONAL_OFFICE_FAX_LITERAL = "rpc_address_fax";
+    public static final String REGIONAL_OFFICE_POSTCODE_LITERAL = "rpc_address_postcode";
+
+    public static final String DWP_ADDRESS_LINE1_LITERAL = "dwp_address_line1";
+    public static final String DWP_ADDRESS_LINE2_LITERAL = "dwp_address_line2";
+    public static final String DWP_ADDRESS_LINE3_LITERAL = "dwp_address_line3";
+    public static final String DWP_ADDRESS_LINE4_LITERAL = "dwp_address_line4";
+
+    static final String ORIGINAL_SENDER_ADDRESS_LINE_1 = "original_sender_address_line1";
+    static final String ORIGINAL_SENDER_ADDRESS_LINE_2 = "original_sender_address_line2";
+    static final String ORIGINAL_SENDER_ADDRESS_LINE_3 = "original_sender_address_line3";
+    static final String ORIGINAL_SENDER_ADDRESS_LINE_4 = "original_sender_address_line4";
+
+    private PlaceholderConstants() {
+        //
+    }
+}
