+++ src/main/java/uk/gov/hmcts/reform/sscs/service/placeholders/RpcPlaceholderService.java
@@ -1,14 +1,14 @@
 package uk.gov.hmcts.reform.sscs.service.placeholders;
 
 import static java.util.Objects.nonNull;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE1_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE2_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE3_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE4_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_COUNTY_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_FAX_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_PHONE_LITERAL;
-import static uk.gov.hmcts.reform.sscs.config.PlaceholderConstants.REGIONAL_OFFICE_POSTCODE_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE1_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE2_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE3_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_ADDRESS_LINE4_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_COUNTY_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_FAX_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_PHONE_LITERAL;
+import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderConstants.REGIONAL_OFFICE_POSTCODE_LITERAL;
 import static uk.gov.hmcts.reform.sscs.service.placeholders.PlaceholderUtility.defaultToEmptyStringIfNull;
 
 import java.util.Map;
@@ -17,9 +17,9 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 
 @Service
-public class RpcPlaceholderService {
+class RpcPlaceholderService {
 
-    public void populatePlaceHolders(Map<String, Object> placeholders, SscsCaseData caseData) {
+    void populatePlaceHolders(Map<String, Object> placeholders, SscsCaseData caseData) {
         if (hasRegionalProcessingCenter(caseData)) {
             RegionalProcessingCenter rpc = caseData.getRegionalProcessingCenter();
             placeholders.put(REGIONAL_OFFICE_ADDRESS_LINE1_LITERAL, defaultToEmptyStringIfNull(rpc.getAddress1()));
